+++ wdt_ct/w8760_funcs.cpp
@@ -619,9 +619,6 @@ int wh_w8760_dev_send_commands(WDT_DEV* pdev, int cmd, UINT32 value)
 				ret = wh_w8760_dev_protect_flash(pdev, W8760_UnprotectLower384k);
 	  		else if (addr < 508 * 1024 && addr + size <= 508 * 1024)
 				ret = wh_w8760_dev_protect_flash(pdev, W8760_UnprotectLower508k);
-		  		else if (addr == 0 && addr + size > 508 * 1024 && addr + size <= 512*1024)
-				ret = wh_w8760_dev_protect_flash(pdev, W8760_UnprotectAll512k);
-
 			return ret;
 		}
 		break;